requirements.txt
openai
pandas
chromadb
python-dotenv
pretty_errors
psycopg2-binary
tiktoken
transformers
tensor
nltk
langdetect
pyarrow
tqdm
scipy